requirements.txt
streamlit>=1.32.0
requests>=2.31.0
pypdf>=4.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
Pillow>=10.0.0
